Support issue synchronization in MQR mode (#1620)
* Add resolve_as_accept

Added the detection and decision whether to synchronize an issue that was resolved as accepted.

The finding is described in this comment in issue #1616 https://github.com/okorach/sonar-tools/issues/1616#issuecomment-2775731968

* Enable issue synchronization in MQR mode

The tool would throw an exception when attempting to synchronize issue status in MQR mode on SonarQube Server. However, the support for this sync is implemented.

Relates to #1616, This comment:
https://github.com/okorach/sonar-tools/issues/1616#issuecomment-2775947372

* Fix issue sync SQC to SQS

* Support issue sync in MQR mode

Fixes #1616.

The 'issueStatus' field has been available since SonarQube Server 10.4 and is the preferred method for retrieving information about issue changes. Starting with SonarQube Server 10.4, the "resolution" and "status" keys are deprecated in issue changelogs but remain relevant for security hotspot changelogs. These conditions are retained for backward compatibility to support versions from SQS 9.9 to 10.3, where "resolution" and "status" are the only way to detect status changes. They are also still applicable for security hotspot changelogs.

* Remove irrelevant section

The docs section about permissions was duplicated, so the irrelevant one was removed.